pyproject.toml
[tool.poetry]
name = "great-reads"
version = "0.1.0"
description = ""
authors = ["rezmotaqi <rezmotaqi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
pycryptodome = "^3.21.0"
motor = "^3.6.0"
jwt = "^1.3.1"
pydantic-settings = "^2.5.2"
aioredis = "^2.0.1"
password-strength = "^0.0.3.post2"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
